created new fabric_node_id.hpp and cpp files
