+++ tt_metal/fabric/CMakeLists.txt
@@ -11,6 +11,7 @@ target_sources(
         fabric.cpp
         fabric_host_utils.cpp
         fabric_context.cpp
+        fabric_node_id.cpp
         serialization/intermesh_link_table.cpp
 )
