+++ tt_metal/impl/dispatch/topology.cpp
@@ -45,6 +45,7 @@
 #include <umd/device/tt_core_coordinates.h>
 #include <umd/device/tt_xy_pair.h>
 #include "utils.hpp"
+#include <tt-metalium/fabric_node_id.hpp>
 
 // hack for test_basic_fabric_apis.cpp
 // https://github.com/tenstorrent/tt-metal/issues/20000
